Projection query refactors
