+++ web/src/views/web/page/index.ejs
@@ -15,7 +15,8 @@
 
         <h2 style="margin-top: 1em;">Utilities</h2>
 
-        <p><a class="button" href="/divider">Division - Personnel Who Work In Every Warehouse</a></p>
+        <a class="button" href="/divider">Division - Personnel Who Work In Every Warehouse</a><br />
+        <a class="button" href="/inventory/projection">Projection - SerialNum, ContainerID, TypeName and Weight from Inventory</a><br />
       </div>
     </main>
