--- web/src/views/web/page/inventory/hard-coded-select.ejs
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <%- include('../../layout/head.ejs') -%>
-</head>
-<body>
-<%- include('../../layout/header.ejs') -%>
-
-<main>
-    <div class="hero">
-        <h1><%= title %></h1>
-        <table>
-            <thead>
-            <tr>
-                <th>Serial Num</th>
-                <th>Container ID</th>
-                <th>Type Name</th>
-                <th>Weight</th>
-                <th></th> <!-- Link/action column -->
-            </tr>
-            </thead>
-            <tbody>
-            <% if (inventoryList) {%>
-                <% inventoryList.forEach(i => { %>
-                    <tr>
-                        <td><%= i.serial_num %></td>
-                        <td><%= i.container_id %></td>
-                        <td><%= i.type_name %></td>
-                        <td><%= i.weight %></td>
-                    </tr>
-                <% }); %>
-            <% } %>
-            </tbody>
-        </table>
-
-    </div>
-</main>
-
-<%- include('../../layout/foot.ejs') -%>
-</body>
-</html>
