+++ web/src/views/web/page/inventory/projection.ejs
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <%- include('../../layout/head.ejs') -%>
+</head>
+<body>
+  <%- include('../../layout/header.ejs') -%>
+
+  <main>
+    <div class="hero">
+      <h1><%= title %></h1>
+      <table>
+        <thead>
+        <tr>
+          <th>Serial Num</th>
+          <th>Container ID</th>
+          <th>Type Name</th>
+          <th>Weight</th>
+        </tr>
+        </thead>
+        <tbody>
+        <% if (inventoryList) {%>
+          <% inventoryList.forEach(i => { %>
+            <tr>
+              <td><%= i.serial_num %></td>
+              <td><%= i.container_id %></td>
+              <td><%= i.type_name %></td>
+              <td><%= i.weight %></td>
+            </tr>
+          <% }); %>
+        <% } %>
+        </tbody>
+      </table>
+
+      <p>This is the result of the following query:</p>
+      <pre>
+  SELECT serial_num, container_id, type_name, weight
+  FROM inventory;</pre>
+
+    </div>
+  </main>
+
+  <%- include('../../layout/foot.ejs') -%>
+</body>
+</html>
